ircd::m::event::prev: Use size_t indexing rather than uint.

+++ include/ircd/m/event/prev.h
@@ -35,13 +35,13 @@ struct ircd::m::event::prev
 {
 	enum cond :int;
 
-	std::tuple<event::id, string_view> auth_events(const uint &idx) const;
-	std::tuple<event::id, string_view> prev_states(const uint &idx) const;
-	std::tuple<event::id, string_view> prev_events(const uint &idx) const;
+	std::tuple<event::id, string_view> auth_events(const size_t &idx) const;
+	std::tuple<event::id, string_view> prev_states(const size_t &idx) const;
+	std::tuple<event::id, string_view> prev_events(const size_t &idx) const;
 
-	event::id auth_event(const uint &idx) const;
-	event::id prev_state(const uint &idx) const;
-	event::id prev_event(const uint &idx) const;
+	event::id auth_event(const size_t &idx) const;
+	event::id prev_state(const size_t &idx) const;
+	event::id prev_event(const size_t &idx) const;
 
 	size_t auth_events_count() const;
 	size_t prev_states_count() const;
